ajout minmax, update freeAVL

+++ c-wire.sh
@@ -85,13 +85,26 @@ fi
 
 
 #Remove and create new tmp and graphs directories. If they exist previously it clean them up.
+if ! [ -d "input" ]; then
+	mkdir input
+fi
+if ! [ -d "tests" ]; then
+	mkdir tests
+fi
+if [ -d "tmp" ]; then
+	rm -rf "tmp"
+fi
+if [ -d "graphs" ]; then
+	rm -rf "graphs"
+fi
 
-rm -rf "tmp"
-rm -rf "graphs"
-mkdir "tmp"
-mkdir "graphs"
+mkdir tmp
+mkdir graphs
+
 rm -f hv*.csv
 rm -f lv*.csv
+
+cp "$1" "input/inputfile.csv"
 
 #Set up the chronometer to mesure execution time of the program
 
@@ -102,10 +115,10 @@ rm -f lv*.csv
 
 
 case $# in 
-	3)	touch "$2_$3.csv"
+	3)	touch "tests/$2_$3.csv"
 		csv_result="$2_$3.csv"
 		echo "$2_$3.csv";;
-	4)	touch "$2_$3_$4.csv"
+	4)	touch "tests/$2_$3_$4.csv"
 		csv_result="$2_$3_$4.csv"
 		echo "$2_$3_$4.csv";;  
 	*) 	;;
@@ -115,25 +128,24 @@ esac
 
 touch tmp/data_to_process.csv
 
-SECONDS="0"
 
 case $2 in
 	#For hva and hvb there is only the comp case
 	hva)
 		if [ $# -eq 4 ]; then
-			grep -E "^$4;[0-9]+;[0-9]+;-;" $1 | cut -d ";" -f3,7,8 | tr "-" "0" >> tmp/data_to_process.csv
+			cat $1 | tail -n+2 | grep -E "^$4;[^;]*;[0-9]+;-;" | cut -d ";" -f3,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 			
 		else
-			grep -E "^[0-9]+;[0-9]+;[0-9]+;-;" $1 | cut -d ";" -f3,7,8 | tr "-" "0" >> tmp/data_to_process.csv
+			cat $1 | tail -n+2 | grep -E "^[0-9]+;[^;]*;[0-9]+;-;" | cut -d ";" -f3,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 			
 		fi;;
  
 	hvb)
 		if [ $# -eq 4 ]; then
-			grep -E "^$4;[0-9]+;-;-;" $1 | cut -d ";" -f2,7,8 | tr "-" "0" >> tmp/data_to_process.csv
+			cat $1 | tail -n+2 | grep -E "^$4;[0-9]+;-;-;" | cut -d ";" -f2,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 			
 		else
-			grep -E "^[0-9]+;[0-9]+;-;-;" $1 | cut -d ";" -f2,7,8 |  tr "-" "0" >> tmp/data_to_process.csv
+			cat $1 | tail -n+2 | grep -E "^[0-9]+;[0-9]+;-;-;" | cut -d ";" -f2,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 			
 		fi;;
 		
@@ -142,67 +154,62 @@ case $2 in
 		case $3 in 
 			comp)
 				if [ $# -eq 4 ]; then
-					grep -E "^$4;-;-;[0-9]+;[0-9]+;-;" $1 | cut -d ";" -f4,7,8 | tr "-" "0" >> tmp/data_to_process.csv
+					cat $1 | tail -n+2 | grep -E "^$4;-;[^;]*;[0-9]+;[^;]*;-;" | cut -d ";" -f4,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 			
 				else
-					grep -E "^[0-9]+;-;-;[0-9]+;[0-9]+;-;" $1 | cut -d ";" -f4,7,8 | tr "-" "0" >> tmp/data_to_process.csv
+					cat $1 | tail -n+2 | grep -E "^[0-9]+;-;[^;]*;[0-9]+;[^;]*;-;" | cut -d ";" -f4,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 					
 				fi;;
 				
 			indiv)
 				if [ $# -eq 4 ]; then
-					grep -E "^$4;-;-;[0-9]+;-;[0-9]+;" $1 | cut -d ";" -f4,7,8 | tr "-" "0" >> tmp/data_to_process.csv
+					cat $1 | tail -n+2 | grep -E "^$4;-;[^;]*;[0-9]+;-;[^;]*;" | cut -d ";" -f4,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 			
 				else
-					grep -E "^[0-9]+;-;-;[0-9]+;-;[0-9]+;" $1 | cut -d ";" -f4,7,8 |  tr "-" "0" >> tmp/data_to_process.csv
+					cat $1 | tail -n+2 | grep -E "^[0-9]+;-;[^;]*;[0-9]+;-;[^;]*;" | cut -d ";" -f4,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 				
 				fi ;;
 			all)
 				if [ $# -eq 4 ]; then
-					grep -E "^$4;-;[^;]*;[0-9]+;" $1 | cut -d ";" -f4,7,8 |  tr "-" "0" >> tmp/data_to_process.csv
+					cat $1 | tail -n+2 | grep -E "^$4;-;[^;]*;[0-9]+;" | cut -d ";" -f4,7,8 | sort -t';' -n -k3  | tr "-" "0" >> tmp/data_to_process.csv
 			
 				else
 				
-					grep -E "^[0-9]+;-;[^;]*;[0-9]+;" $1 | cut -d ";" -f4,7,8 | tr "-" "0" >> tmp/data_to_process.csv
+					cat $1 | tail -n+2 | grep -E "^[0-9]+;-;[^;]*;[0-9]+;" | cut -d ";" -f4,7,8 | sort -t';' -n -k3 | tr "-" "0" >> tmp/data_to_process.csv
 					
 				fi;;
 		esac;; #End of lv case
 	*);;
 esac  
 
-echo "Program's execution time : $SECONDS sec"
 
 touch tmp/data_processed.csv
-
-
-if [ $# -eq 4 ]; then 
-	cat tmp/data_to_process.csv >> ./main "tmp/data_processed.csv"
-	
-else
-	cat tmp/data_to_process.csv >> ./main "tmp/data_processed.csv"
-fi
-
 chmod 777 tmp/data_processed.csv
 
 
 
 #Create the final file
+#SUPPLEMENTARY STEPS TO DO LATER WHEN C PROGRAM IS FINISHED Line 239
 
+cd codeC
+make 
+datapath="tmp/data_to_process.csv"
+./main "../$datapath" "../tmp/data_processed.csv"
 
-
+cd ..
 
 case $2 in
 
 	hva)
 		if [ $# -eq 4 ]; then
 		
-		echo "HV-A;Capacity;Load" >> $2_$3_$4.csv
-		cat tmp/data_processed.csv | grep -E "^[0-9]+;[0-9]+;0" | cut -d ";" -f1,2,4 | sort -t';' -n -k3 >> $2_$3_$4.csv
+		echo "HV-A:Capacity:Load" > "tests/$2_$3_$4.csv"
+		sort tmp/data_processed.csv -t':' -n -k2 >> "tests/$2_$3_$4.csv"
 		
 		else 
 		
-		echo "HV-A;Capacity;Load" >> $2_$3.csv
-		cat tmp/data_processed.csv | grep -E "^[0-9]+;[0-9]+;0" | cut -d ";" -f1,2,4 | sort -t';' -n -k3 >> $2_$3.csv
+		echo "HV-A:Capacity:Load" > "tests/$2_$3.csv"
+		sort tmp/data_processed.csv -t':' -n -k2 >> "tests/$2_$3.csv"
 		
 		fi;;
 	
@@ -211,93 +218,61 @@ case $2 in
 	
 		if [ $# -eq 4 ]; then
 		
-			echo "HV-B;Capacity;Load" >> $2_$3_$4.csv
-			cat tmp/data_processed.csv | grep -E "^[0-9]+;[0-9]+;0" | cut -d ";" -f1,2,4 | sort -t';' -n -k3 > $2_$3_$4.csv
+		echo "HV-B:Capacity:Load" > "tests/$2_$3_$4.csv"
+		sort tmp/data_processed.csv -t':' -n -k2 >> "tests/$2_$3_$4.csv"
 		
 		else 
 	
-			echo "HV-B;Capacity;Load" >> $2_$3.csv
-			cat tmp/data_processed.csv | grep -E "^[0-9]+;[0-9]+;0" | cut -d ";" -f1,2,4 | sort -t';' -n -k3 > $2_$3.csv
+		echo "HV-B:Capacity:Load" > "tests/$2_$3.csv"
+		sort "tmp/data_processed.csv" -t':' -n -k2 >> "tests/$2_$3.csv"
 		
 		fi;;
 		
-	lv) #SUPPLEMENTARY STEPS TO DO LATER WHEN C PROGRAM IS FINISHED
+	lv) 
+		
 		if [ $# -eq 4 ]; then
 		
-			echo "LV;Capacity;Load" >> $2_$3_$4.csv
-			cat tmp/data_processed.csv | grep -E "^[0-9]+;[0-9]+;0" | cut -d ";" -f1,2,4 | sort -t';' -n -k3 > $2_$3_$4.csv
-		
-			if  [ $3 == "all" ]; then 
-				
-				cat $2_$3_$4.csv | head -11 > $2_$3_$4_minmax.csv 
-				cat $2_$3_$4.csv | tail -10 >> $2_$3_$4_minmax.csv 
-				
-				gnuplot -persist <<-EOFMarker
-					gnuplot -persist <<-EOFMarker
-					set xlabel "LV"
-					set ylabel "Capacity/Load"
-					set grid
-					set datafile separator ";"
-					set boxwidth 0.4
-					set style data histogram
-					set style fill solid
-					plot "test.csv" using 1:2 with boxes linecolor rgb "#00FF00", \
-						 "test.csv" using 1:3 with boxes linecolor rgb "#FF0000"
-				EOFMarker
-				
-			fi
+		echo "LV:Capacity:Load" > "tests/$2_$3_$4.csv"
+		sort "tmp/data_processed.csv" -t':' -n -k2 >> "tests/$2_$3_$4.csv" 
+		fileLV="$2_$3_$4" 
 		
 		else 
 	
-			echo "LV;Capacity;Load" >> $2_$3.csv
-			cat tmp/data_processed.csv | grep -E "^[0-9]+;[0-9]+;0" | cut -d ";" -f1,2,4 | sort -t';' -n -k3 >> $2_$3.csv
-		
-			if  [ $3 == "all" ]; then 
-				
-				cat $2_$3.csv | head -11 > $2_$3_minmax.csv 
-				cat $2_$3.csv | tail -10 >> $2_$3_minmax.csv 
-				
-				
-				#Bar graphic
-				
-				gnuplot -persist <<-EOFMarker
-					gnuplot -persist <<-EOFMarker
-					set xlabel "LV"
-					set ylabel "Capacity/Load"
-					set grid
-					set datafile separator ";"
-					set boxwidth 0.4
-					set style data histogram
-					set style fill solid
-					plot "test.csv" using 1:2 with boxes linecolor rgb "#00FF00", \
-						 "test.csv" using 1:3 with boxes linecolor rgb "#FF0000"
-				EOFMarker
+		echo "LV:Capacity:Load" > "tests/$2_$3.csv"
+		sort "tmp/data_processed.csv" -t':' -n -k2 >> "tests/$2_$3.csv"
+		fileLV="$2_$3" 
 
+		fi
+
+		if [ "$3" == "all" ]; then
+			tail -n+2 "tests/$fileLV.csv" | sort -t':' -n -k3 > "tmp/sortload.csv"
+			head -n10 "tmp/sortload.csv" >> "tmp/sortdiff.csv"
+			tail -n10 "tmp/sortload.csv" >> "tmp/sortdiff.csv"
+			echo "LV:Capacity:Load:Energy Difference" > tests/"$fileLV"_minmax.csv
+			awk -F: '{$4=$2-$3} {print $0}' OFS=: "tmp/sortdiff.csv" | sort -t':' -n -k4 >> tests/"$fileLV"_minmax.csv
 			
-				sort -t ';' -n -k3 $2_$3_minmax.csv 
-			fi
-		
+			#Bar graphic
+				
+			#	gnuplot -persist <<-EOFMarker
+			#		gnuplot -persist <<-EOFMarker
+			#		set xlabel "LV"
+			#		set ylabel "Capacity/Load"
+			#		set grid
+			#		set datafile separator ";"
+			#		set boxwidth 0.4
+			#		set style data histogram
+			#		set style fill solid
+			#		plot "test.csv" using 1:2 with boxes linecolor rgb "#00FF00", \
+			#			 "test.csv" using 1:3 with boxes linecolor rgb "#FF0000"
+			#	EOFMarker
+			
+
 		fi;;
+
 	*);;
-
 esac 
-
-
-
-
-cd codeC
-gcc -o main CreationArbreAVL.c 
-./main "../$datapath" "$2_$3.csv"
-
-
-
 
 
 #print program's execution time
 
 echo "Program's execution time : $SECONDS sec"
-
-
-
-
-
